requirements.txt
pandas
biopython
scikit-learn
streamlit
umap-learn
hdbscan
matplotlib
